requirements.txt
bcrypt==4.1.3
fastapi==0.110.0
passlib==1.7.4
pydantic_core==2.18.4
pydantic-mongo==2.3.0
pydantic-settings==2.3.4
PyJWT==2.8.0
pymongo==4.6.2
pypdf==3.17.4
clean-text==0.6.0
aiofiles==24.1.0
reportlab==4.2.2
uvicorn[standard]==0.30.1
python-multipart==0.0.9
openai==1.43.0
